+++ apps/e2e/layout-storybook-e2e/src/e2e/action-button.component.cy.ts
@@ -2,6 +2,8 @@ import { E2eVariations } from '@skyux-sdk/e2e-schematics';
 
 describe('layout-storybook - action button', () => {
   E2eVariations.forEachTheme((theme) => {
+    const screenshotPrefix = `actionbuttoncomponent-actionbutton--action-button-${theme}`;
+
     describe(`in ${theme} theme`, () => {
       beforeEach(() =>
         cy.visit(
@@ -12,31 +14,46 @@ describe('layout-storybook - action button', () => {
         cy.get('app-action-button')
           .should('exist')
           .should('be.visible')
-          .screenshot(
-            `actionbuttoncomponent-actionbutton--action-button-${theme}`
-          )
-          .percySnapshot(
-            `actionbuttoncomponent-actionbutton--action-button-${theme}`,
-            {
-              widths: E2eVariations.DISPLAY_WIDTHS,
-            }
-          );
+          .screenshot(screenshotPrefix)
+          .percySnapshot(screenshotPrefix, {
+            widths: E2eVariations.DISPLAY_WIDTHS,
+          });
+      });
+
+      it('should render in a medium modal', () => {
+        const screenshotName = `${screenshotPrefix}-modal`;
+
+        cy.get('app-action-button')
+          .should('exist')
+          .should('be.visible')
+          .get('.action-button-open-modal')
+          .should('exist')
+          .should('be.visible')
+          .click()
+          .end()
+          .get('.sky-modal')
+          .should('exist')
+          .should('be.visible')
+          .screenshot(screenshotName, {
+            disableTimersAndAnimations: true,
+            scale: false,
+          })
+          .percySnapshot(screenshotName, {
+            widths: E2eVariations.DISPLAY_WIDTHS,
+          });
       });
 
       it('should render the component on mobile', () => {
+        const screenshotName = `${screenshotPrefix}-mobile`;
+
         cy.viewport(E2eVariations.MOBILE_WIDTHS[0], 800);
         cy.get('app-action-button')
           .should('exist')
           .should('be.visible')
-          .screenshot(
-            `actionbuttoncomponent-actionbutton--action-button-${theme}-mobile`
-          )
-          .percySnapshot(
-            `actionbuttoncomponent-actionbutton--action-button-${theme}-mobile`,
-            {
-              widths: E2eVariations.MOBILE_WIDTHS,
-            }
-          );
+          .screenshot(screenshotName)
+          .percySnapshot(screenshotName, {
+            widths: E2eVariations.MOBILE_WIDTHS,
+          });
       });
     });
   });