+++ libs/components/layout/src/lib/modules/toolbar/toolbar.component.ts
@@ -29,12 +29,14 @@ export class SkyToolbarComponent {
   }
 
   /**
-   * A descriptor for the items that the toolbar manipulates. Use a plural term. The descriptor helps set the toolbar’s `aria-label` attributes for search inputs, sort buttons, and filter buttons to provide text equivalents for screen readers [to support accessibility](https://developer.blackbaud.com/skyux/components/checkbox#accessibility).
-   * For example, when the descriptor is “constituents,” the search input’s `aria-label` is “Search constituents.” For more information about the `aria-label` attribute, see the [WAI-ARIA definition](https://www.w3.org/TR/wai-aria/#aria-label).
+   * A descriptor for the items that the toolbar manipulates. Use a plural term. The descriptor helps set the toolbar's `aria-label` attributes for search inputs, sort buttons, and filter buttons to provide text equivalents for screen readers [to support accessibility](https://developer.blackbaud.com/skyux/components/checkbox#accessibility).
+   * For example, when the descriptor is "constituents," the search input's `aria-label` is "Search constituents." For more information about the `aria-label` attribute, see the [WAI-ARIA definition](https://www.w3.org/TR/wai-aria/#aria-label).
    */
   @Input()
   public set listDescriptor(value: string | undefined) {
-    this.#contentInfoProvider.patchInfo({ descriptor: value });
+    this.#contentInfoProvider.patchInfo({
+      descriptor: value ? { type: 'text', value } : undefined,
+    });
   }
 
   #contentInfoProvider = inject(SkyContentInfoProvider);